ligação da página de posts com comentários
