+++ projeto-4eddit8/debug.log
@@ -40,3 +40,17 @@
 [0710/112852.559:ERROR:crash_report_database_win.cc(469)] failed to stat report
 [0710/112852.559:ERROR:crash_report_database_win.cc(469)] failed to stat report
 [0710/112852.559:ERROR:crash_report_database_win.cc(469)] failed to stat report
+[0710/145300.825:ERROR:crash_report_database_win.cc(469)] failed to stat report
+[0710/145300.825:ERROR:crash_report_database_win.cc(469)] failed to stat report
+[0710/145300.825:ERROR:crash_report_database_win.cc(469)] failed to stat report
+[0710/145300.825:ERROR:crash_report_database_win.cc(469)] failed to stat report
+[0710/145300.826:ERROR:crash_report_database_win.cc(469)] failed to stat report
+[0710/145300.826:ERROR:crash_report_database_win.cc(469)] failed to stat report
+[0710/145300.826:ERROR:crash_report_database_win.cc(469)] failed to stat report
+[0710/145300.826:ERROR:crash_report_database_win.cc(469)] failed to stat report
+[0710/145300.826:ERROR:crash_report_database_win.cc(469)] failed to stat report
+[0710/145300.826:ERROR:crash_report_database_win.cc(469)] failed to stat report
+[0710/145300.826:ERROR:crash_report_database_win.cc(469)] failed to stat report
+[0710/145300.826:ERROR:crash_report_database_win.cc(469)] failed to stat report
+[0710/145300.826:ERROR:crash_report_database_win.cc(469)] failed to stat report
+[0710/145300.826:ERROR:crash_report_database_win.cc(469)] failed to stat report
